Deprecate Zeus Modules in favor of clearer UI
Previously, share interactions would still be shown even when the restriction was disabled, in order to allow local markers to be shared if the restriction was lifted mid-mission via Zeus Modules.
In favor of keeping the UI clearer by not showing the interactions, the Zeus Modules were removed (their use case was moot anyway).