More restrictive "player created" classification
Mod and Mission Scripts often add "_USER_DEFINED" in their marker names, so that they will be editable by players interacting with the map. The issue is that they would then be recognized as player-created by Restrict Markers.
Now, thanks to a more restrictive classification, those scripts can add a suffix to their marker names, so that they won't be counted as "player created" anymore.